Refactor: Extract facet-mixin
